ImageCaption/blip-image-captioning-api/requirements.txt
fastapi
uvicorn
transformers
pillow
cython
torch
pydantic
pytest
requests
pydantic-settings
PyYAML
python-multipart
streamlit
Pillow